Uppercase a3m files
Otherwise you get the following error:

setting an array element with a sequence.
TypeError: int() argument must be a string, a bytes-like object or a number, not 'list'

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/home/konsti/ColabFold/colabfold/batch.py", line 506, in run
    **pipeline.make_msa_features(
  File "/home/konsti/ColabFold/.venv/lib/python3.8/site-packages/alphafold/data/pipeline.py", line 74, in make_msa_features
    features['deletion_matrix_int'] = np.array(deletion_matrix, dtype=np.int32)
ValueError: setting an array element with a sequence.